Add codespell line suppression file
This file is used with "-x misc/suppressions/codespell-lines" codespell
option and allows to ignore entire lines, instead of ignoring words that
only occur once or twice and shouldn't be ignored globally because they
could well be misspelt elsewhere.

The remaining words in codespell.ignore occur too many times (as
parameter names in various places) to be ignored in this way.