Don't scroll when inserting items into wxListBox in wxOSX
This seems to have been needed in the past (see #12365), but is not
needed any more and results in unexpectedly scrolling down to the last
inserted item, which is undesirable and inconsistent with the other
platforms.

Simply revert 07bee39fb6 (Show the first, not the last, inserted item in
wxListBox in wxOSX., 2010-11-05) to fix this.

Closes #18861.